+++ NEWS.md
@@ -48,6 +48,8 @@
 
 * `ddbs_quadkey()`: calculates quadkey tiles from point geometries
 
+* `ddbs_exterior_ring()`: returns the exterior ring (shell) of a polygon geometry
+
 * **Spatial predicates**: spatial predicates are all included in a function called `ddbs_predicate()`, where the user can specify the spatial predicate. Another option, it's to use the spatial predicate function, such as `ddbs_intersects()`, `ddbs_crosses()`, `ddbs_touches()`, etc.
 
 ## MINOR CHANGES